Fixed alerts

+++ PyGLM/type_methods/glmArray.h
@@ -1291,7 +1291,7 @@ PyBuffer_Release(&view);\
 return (PyObject*)out
 
 #define GLM_ARRAY_AS_REFERENCE_IF_IS_MAT_BUFFER(T) \
-out->itemSize = C * R * out->dtSize;\
+out->itemSize = (ssize_t)C * (ssize_t)R * out->dtSize;\
 out->subtype = (C == 2) ? (R == 2) ? PyGLM_MAT_TYPE<2, 2, T>() : (R == 3) ? PyGLM_MAT_TYPE<2, 3, T>() : PyGLM_MAT_TYPE<2, 4, T>() :\
 (C == 3) ? (R == 2) ? PyGLM_MAT_TYPE<3, 2, T>() : (R == 3) ? PyGLM_MAT_TYPE<3, 3, T>() : PyGLM_MAT_TYPE<3, 4, T>() :\
 (R == 2) ? PyGLM_MAT_TYPE<4, 2, T>() : (R == 3) ? PyGLM_MAT_TYPE<4, 3, T>() : PyGLM_MAT_TYPE<4, 4, T>();\
